Add links to gpl source
I just realise the hardest part of this project is wifi driver. I can make adapter connect to other but make itself to be an AP is different story. It also need driver to support.

So far, I can make an AP from RT5572 adapter (Dlink WDA-160) but performance is quiet bad compare to ethernet.